requirements.txt
# ── Core ──────────────────────────────────────────────────────────────────────
Django==4.2.16
djangorestframework==3.15.2
django-allauth==0.63.6
whitenoise==6.7.0

# ── Database ──────────────────────────────────────────────────────────────────
psycopg2-binary==2.9.9
dj-database-url==2.2.0

# ── Config ────────────────────────────────────────────────────────────────────
python-decouple==3.8

# ── Auth (allauth JWT dependency) ─────────────────────────────────────────────
PyJWT>=2.8.0
cryptography>=42.0.0

# ── AI ────────────────────────────────────────────────────────────────────────
google-genai>=1.0.0

# ── Payments ──────────────────────────────────────────────────────────────────
stripe==10.7.0

# ── Document parsing ──────────────────────────────────────────────────────────
PyMuPDF==1.24.9
pdfplumber>=0.11.0
python-docx==1.1.2

# ── Server ────────────────────────────────────────────────────────────────────
gunicorn==22.0.0

# ── Pillow (optional, for future image handling) ─────────────────────────────
Pillow==10.4.0
